requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.10
redis==5.2.1
xlrd==2.0.1
odfpy==1.4.1
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
email-validator==2.2.0
